Formatted Related Item cards so they properly display live data, modal button changes color on hover, and created a file for helper function for the server.
Inside the helper function right now is just a function that takes in an array of ratings and finds the average rating to be used to calculate stars.

I am making axios calls inside my parent component for now, but eventually those calls should be moved to our server file.